requirements.txt
selenium
flask
pyyaml
